fixed inital deposit

+++ ATM_GUI.py
@@ -416,8 +416,8 @@ def checkEntryBoxes():
         messagebox.showerror("Invalid input", "You cannot leave \"Confirm PIN\" empty")
     else:
         try:
-            int(deposit.get())
-            if(int(deposit.get()) <0 or int(deposit.get()) > ATM.MAX_BALANCE):
+            float(deposit.get())
+            if(float(deposit.get()) < 0 or float(deposit.get()) > ATM.MAX_BALANCE):
                 messagebox.showerror("Invalid input for initial deposit",
                                  "Your deposit must be an integer that is >0 and <999999999999")
             else:
